Added tests

+++ src/main/java/com/joe/benefits/employee/model/EmployeePaycheck.java
@@ -4,6 +4,8 @@ import jakarta.persistence.*;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @Table(name = "EMPLOYEE_PAYCHECK")
@@ -19,9 +21,9 @@ public class EmployeePaycheck {
     @JoinColumn(name = "BENEFIT_ID")
     private BenefitPackage benefitPackage;
     @Column(name = "TOTAL")
-    private Double total;
+    private BigDecimal total;
     @Column(name = "SALARY")
-    private Double salary;
+    private BigDecimal salary;
     @Column(name = "status")
     @Enumerated(EnumType.STRING)
     private PayrollStatus status;
